requirements2.txt
<<<<<<< HEAD
# 기본 패키지
torch==2.1.0
torchvision==0.16.0
pytorch-lightning==2.5.2
hydra-core==1.3.2
omegaconf==2.3.0

# 데이터 증강
albumentations==1.3.1
augraphy==8.2.6
opencv-python>=4.5.0

# metrics
torchmetrics==1.2.0

# 실험 관리
wandb==0.21.0
=======
# ⚠️ 시스템 의존성: OpenCV 사용을 위해 아래 라이브러리 필요
# Ubuntu 기준: sudo apt-get install -y libglib2.0-0

wandb==0.21.0
pytorch-lightning==2.5.2
hydra-core==1.3.2
omegaconf==2.3.0
torchmetrics==1.3.2
augraphy==8.2.6
opencv-python==4.7.0.72
>>>>>>> 9c70dc66e655ab4673040036548c13ec65447c8b
